+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- The maximum buffer length of the "Digest Update Multiple" SAM command is set to 230 bytes instead of the 254 
+  theoretically allowed by the SAM in order to be compatible with certain unpredictable applications.
 
 ## [2.2.5] - 2022-10-17
 ### Fixed