+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.0] - 2022-12-06
 ### Added
 - `ContextSetting` setting to `CalypsoExtensionService` to manage the limitations of some not fully compliant terminals.
 ### Fixed
@@ -126,7 +128,8 @@ This is the initial release.
 It follows the extraction of Keyple 1.0 components contained in the `eclipse/keyple-java` repository to dedicated repositories.
 It also brings many major API changes.
 
-[unreleased]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.2.5...HEAD
+[unreleased]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.3.0...HEAD
+[2.3.0]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.2.5...2.3.0
 [2.2.5]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.2.4...2.2.5
 [2.2.4]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.2.2...2.2.3